feat(client): add pagination information to service index methods
BREAKING CHANGES:

The index and search methods of all services now returns both the data
and any provided pagination information, so instead of e.g. an array, it
will now be an object containing data.

    interface Index<T> {
      data: Array<T>;
      page: Page;
    }

BEFORE:

    // Observable<Array<Foo>>
    // [...]
    this.fooService.index({ limit: 5 })

AFTER:

    // Observable<Index<Foo>>
    // { data: [...], page: { ... } }
    this.fooService.index({ limit: 5 })

---

There are a number of features, planned for Gear and implemented in
Personnel, which require the pagination data for an index method. This
commit changes index and search methods to return pagination alongside
the data.

- API
  - Add Index<T> for index and search methods.
- Services
  - Change the index and search methods of all services to return
    pagination and array data.
  - Update and standardize syntax used in all services. All services
    should now look and behave in an identical manner.
- Effects
  - Update all calling effects in Gear.
- Testing
  - Add testing factory for pagination data.
  - Update all tests.

Closes: d4h/decisions-project#4224